Rename to uberComponents/ResourceListPage
