+++ vite.lib.config.ts
@@ -37,6 +37,7 @@ export default defineConfig({
             entry: [
                 path.resolve(__dirname, 'src/index.ts'),
                 path.resolve(__dirname, 'src/components/index.ts'),
+                path.resolve(__dirname, 'src/uberComponents/index.ts'),
                 path.resolve(__dirname, 'src/containers/index.ts'),
                 path.resolve(__dirname, 'src/hooks/index.ts'),
                 path.resolve(__dirname, 'src/utils/index.ts'),
